Modify access.php config. Duplicate afisha routing for compability with older links. Add ajax access to backend. Fix bug Sidebar for seo links. Change backend views (add right rules).

+++ backend/Views/Afisha/Menu.php
@@ -25,13 +25,17 @@
 		                        <li>
 		                            <a class="bs-tooltip dropdownToggle" href="javascript:void(0);" title="Управление"><i class="fa-cog size14"></i></a>
 		                            <ul class="dropdownMenu pull-right">
-		                                <li>
-		                                    <a href="/backend/<?php echo \Core\Route::controller(); ?>/edit/<?php echo $obj->id; ?>" title="Редактировать"><i class="fa-pencil"></i> Редактировать</a>
-		                                </li>
-		                                <li class="divider"></li>
-		                                <li>
-		                                    <a onclick="return confirm('Это действие необратимо. Продолжить?');" href="/backend/<?php echo \Core\Route::controller(); ?>/delete/<?php echo $obj->id; ?>" title="Удалить"><i class="fa-trash-o text-danger"></i> Удалить</a>
-		                                </li>
+										<?php if(Core\User::caccess() == 'edit' OR Core\User::caccess() == 'view'): ?>
+											<li>
+												<a href="/backend/<?php echo \Core\Route::controller(); ?>/edit/<?php echo $obj->id; ?>" title="Редактировать"><i class="fa-pencil"></i> Редактировать</a>
+											</li>
+										<?php endif; ?>
+										<?php if(Core\User::caccess() == 'edit'): ?>
+											<li class="divider"></li>
+											<li>
+												<a onclick="return confirm('Это действие необратимо. Продолжить?');" href="/backend/<?php echo \Core\Route::controller(); ?>/delete/<?php echo $obj->id; ?>" title="Удалить"><i class="fa-trash-o text-danger"></i> Удалить</a>
+											</li>
+										<?php endif; ?>
 		                            </ul>
 		                        </li>
 		                    </ul>
